Renames GKE bucket backup container
Two name changes:

  1. container image: "gs_bucket_backup" -> "gke-gs-bucket-backup"
  2. script: "bucket_sync.py" -> "gs_bucket_sync.py"

Using a prefix "gke-" on the container name, since we'll likely have a
handful of these helpers in the near future.
