+++ gke-gs-bucket-backup/.dockerignore
