+++ gke-gs-bucket-backup/Dockerfile
@@ -3,11 +3,11 @@ ARG CLOUD_SDK_HASH='d026fcb44de9f3ac58ed959afa892d8216de858dc69b370e001d641a4e36
 
 FROM google/cloud-sdk@sha256:${CLOUD_SDK_HASH}
 
-COPY bucket_sync.py /usr/local/bin
+COPY gs_bucket_sync.py /usr/local/bin
 
 RUN adduser --disabled-password --gecos "" gcloud_user
 USER gcloud_user
 ENV PATH "$PATH:/opt/google-cloud-sdk/bin/"
 
 
-ENTRYPOINT [ "/usr/local/bin/bucket_sync.py" ]
+ENTRYPOINT [ "/usr/local/bin/gs_bucket_sync.py" ]
